#17: Stub implementations moved into separate subpackage
