+++ src/test/scala/org/fs/mael/backend/http/HttpBackendSpec.scala
@@ -2,16 +2,16 @@ package org.fs.mael.backend.http
 
 import java.net.URI
 
-import org.fs.mael.test.StubEventManager
+import org.fs.mael.core.transfer.SimpleTransferManager
+import org.fs.mael.test.stub.StoringEventManager
 import org.junit.runner.RunWith
 import org.scalatest.FunSuite
-import org.fs.mael.core.transfer.SimpleTransferManager
 
 @RunWith(classOf[org.scalatest.junit.JUnitRunner])
 class HttpBackendSpec
   extends FunSuite {
 
-  val backend = new HttpBackend(new StubEventManager, new SimpleTransferManager)
+  val backend = new HttpBackend(new StoringEventManager, new SimpleTransferManager)
 
   test("supported URLs") {
     assert(backend.isSupported(new URI("http://abcde")))
