+++ src/test/scala/org/fs/mael/core/event/EventManagerImplSpec.scala
@@ -6,7 +6,7 @@ import java.net.URI
 import org.fs.mael.core.Status
 import org.fs.mael.core.entry.LogEntry
 import org.fs.mael.core.event.Events._
-import org.fs.mael.test.StubBackend
+import org.fs.mael.test.stub.StubBackend
 import org.junit.runner.RunWith
 import org.scalatest.BeforeAndAfter
 import org.scalatest.FunSuite
