+++ src/test/scala/org/fs/mael/test/stub/StoringEventManager.scala
@@ -1,21 +1,25 @@
-package org.fs.mael.test
+package org.fs.mael.test.stub
 
 import org.fs.mael.core.event.EventManager
 import org.fs.mael.core.event.EventSubscriber
 import org.fs.mael.core.event.PriorityEvent
 
 /** Event manager that accumulates events */
-class StubEventManager extends EventManager {
+class StoringEventManager extends EventManager {
   var events: IndexedSeq[PriorityEvent] = IndexedSeq.empty
 
-  override def subscribe(subscriber: EventSubscriber): Unit = { /* NOOP */ }
+  override def subscribe(subscriber: EventSubscriber): Unit = {
+    /* NOOP */
+  }
 
-  override def unsubscribe(id: String): Unit = { /* NOOP */ }
+  override def unsubscribe(id: String): Unit = {
+    /* NOOP */
+  }
 
   override def fire(event: PriorityEvent): Unit = {
     events = events :+ event
   }
 
-  def clear(): Unit =
+  def reset(): Unit =
     events = IndexedSeq.empty
 }
